safari-534.53-branch/LayoutTests/sputnik/Conformance/12_Statement/12.14_The_try_Statement/S12.14_A16_T8-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Parse error
S12.14_A16_T8

PASS Expected parsing failure

TEST COMPLETE
